Minor clarification
Hey, I burned an hour on this because it would sometimes enqueue the record correctly, so I figured it was a multithread bug of some kind when in reality it was just a class var and a race condition. 

Feel free to decline the PR if you want to keep the README small, but I think it's helpful to point out.